File Transfer
Summary: Transfer data at the same time so data isn't missed. 

This means you will have to provide for recording the last time the 'file check' process was performed,
so that you can be sure to cover the entire time period in which new or edited files could occur.
To do this, you will need to create a database with a table that can store the date and time of the last 'file
check' process. That way, you can use that date/time as a reference point in terms of finding new or
modified files.
As part of this project, the users are asking that their UI display the date and time of the last 'file check'
process.
You have been asked to implement this functionality. This means that you will need to
• create a database and a table
• modify your script to both record date/time of 'file check' runs and to retrieve that data for use in
the 'file check' process, and
• modify the UI to display the last 'file check' date/time
Guidelines:
Use Python 3.4 for this drill.
Use sqlite for your database functionality.
Use tkinter module for the UI.